Credit_scores_project
The project is to prepare a report for a bank’s loan division. will need to find out if a customer’s marital status and number of children have an impact on whether they will default on a loan. The bank already has some data on customers’ credit worthiness.
The report will be considered when building a credit score for a potential customer. a credit score is used to evaluate the ability of a potential borrower to repay their loan.