market_research_project
opening a small robot-run cafe in Los Angeles. the project is promising but expensive, so partners decided to try to attract investors. they’re interested in the current market conditions—will you be able to maintain your success when the novelty of robot waiters wears off?
partners have asked you to prepare some market research. we have open-source data on restaurants in LA.