users_behavior_project
first study the sales funnel. will find out how users reach the purchase stage. how many users actually make it to this stage? how many get stuck at previous stages? which stages in particular? The designers would like to change the fonts for the entire app, but the managers are afraid the users might find the new design intimidating. they decide to make a decision based on the results of an A/A/B test.